Refactor staking module
